XRT-623 Move struct cma_bank to userpf parent driver (#3644)
