+++ src/runtime_src/core/pcie/driver/linux/xocl/userpf/xocl_drm.c
@@ -598,55 +598,6 @@ uint32_t xocl_get_shared_ddr(struct xocl_drm *drm_p, struct mem_data *m_data)
 	return 0xffffffff;
 }
 
-static void xocl_cma_mem_free(struct xocl_drm *drm_p, uint32_t idx)
-{
-	struct xocl_cma_memory *cma_mem = &drm_p->cma_bank->cma_mem[idx];
-	struct sg_table *sgt = NULL;
-
-	if (!cma_mem)
-		return;
-
-	sgt = cma_mem->sgt;
-	if (sgt) {
-		dma_unmap_sg(drm_p->ddev->dev, sgt->sgl, sgt->orig_nents, DMA_BIDIRECTIONAL);
-		sg_free_table(sgt);
-		vfree(sgt);
-		cma_mem->sgt = NULL;
-	}
-
-	if (cma_mem->vaddr) {
-		dma_free_coherent(&drm_p->ddev->pdev->dev, cma_mem->size, cma_mem->vaddr, cma_mem->paddr);
-		cma_mem->vaddr = NULL;
-	} else if (cma_mem->pages) {
-#if LINUX_VERSION_CODE >= KERNEL_VERSION(4, 15, 0)
-		release_pages(cma_mem->pages, cma_mem->size >> PAGE_SHIFT);
-#else
-		release_pages(cma_mem->pages, cma_mem->size >> PAGE_SHIFT, 0);
-#endif
-	}
-
-	if (cma_mem->pages) {
-		vfree(cma_mem->pages);
-		cma_mem->pages = NULL;
-	}
-}
-
-static void xocl_cma_mem_free_all(struct xocl_drm *drm_p)
-{
-	int i = 0;
-	uint64_t num = 0;
-
-	if (!drm_p->cma_bank)
-		return;
-
-	num = drm_p->cma_bank->entry_num;
-
-	for (i = 0; i < num; ++i)
-		xocl_cma_mem_free(drm_p, i);
-
-	xocl_info(drm_p->ddev->dev, "%s done", __func__);
-}
-
 int xocl_cleanup_mem_nolock(struct xocl_drm *drm_p)
 {
 	int err;
@@ -908,382 +859,4 @@ bool is_cma_bank(struct xocl_drm *drm_p, uint32_t memidx)
 done:
 	XOCL_PUT_MEM_TOPOLOGY(drm_p->xdev);
 	return ret;
-}
-
-static int xocl_cma_mem_alloc_huge_page_by_idx(struct xocl_drm *drm_p, uint32_t idx, uint64_t user_addr, uint64_t page_sz)
-{
-	uint64_t page_count = 0, nr = 0;
-	struct device *dev = drm_p->ddev->dev;
-	int ret = 0;
-	struct xocl_cma_memory *cma_mem = &drm_p->cma_bank->cma_mem[idx];
-	struct sg_table *sgt = NULL;
-
-	if (!(XOCL_ACCESS_OK(VERIFY_WRITE, user_addr, page_sz))) {
-		xocl_err(dev, "Invalid huge page user pointer\n");
-		ret = -ENOMEM;
-		goto done;
-	}
-
-	page_count = (page_sz) >> PAGE_SHIFT;
-	cma_mem->pages = vzalloc(page_count*sizeof(struct page *));
-	if (!cma_mem->pages) {
-		ret = -ENOMEM;
-		goto done;
-	}
-
-	nr = get_user_pages_fast(user_addr, page_count, 1, cma_mem->pages);
-	if (nr != page_count) {
-		xocl_err(dev, "Can't pin down enough page_nr %llx\n", nr);
-		ret = -EINVAL;
-		goto done;
-	}
-
-	sgt = vzalloc(sizeof(struct sg_table));
-	if (!sgt) {
-		ret = -ENOMEM;
-		goto done;
-	}
-
-	ret = sg_alloc_table_from_pages(sgt, cma_mem->pages, page_count, 0, page_sz, GFP_KERNEL);
-	if (ret) {
-		ret = -ENOMEM;
-		goto done;
-	}
-
-	if (sgt->orig_nents != 1) {
-		xocl_err(dev, "Host mem is not physically contiguous\n");
-		ret = -EINVAL;
-		goto done;
-	}
-
-	if (!dma_map_sg(dev, sgt->sgl, sgt->orig_nents, DMA_BIDIRECTIONAL)) {
-		ret =-ENOMEM;
-		goto done;
-	}
-
-	if (sgt->orig_nents != sgt->nents) {
-		ret =-ENOMEM;
-		goto done;		
-	}
-
-	cma_mem->size = page_sz;
-	cma_mem->paddr = sg_dma_address(sgt->sgl);
-	cma_mem->sgt = sgt;
-
-done:
-	if (ret) {
-		vfree(cma_mem->pages);
-		cma_mem->pages = NULL;
-		if (sgt) {
-			dma_unmap_sg(dev, sgt->sgl, sgt->orig_nents, DMA_BIDIRECTIONAL);
-			sg_free_table(sgt);
-			vfree(sgt);
-		}
-	}
-
-	return ret;
-}
-
-static int xocl_cma_mem_alloc_huge_page(struct xocl_drm *drm_p, struct drm_xocl_alloc_cma_info *cma_info)
-{
-	int ret = 0;
-	xdev_handle_t xdev = drm_p->xdev;
-	size_t page_sz = cma_info->total_size/cma_info->entry_num;
-	uint32_t i, j, num = xocl_addr_translator_get_entries_num(xdev);
-	uint64_t *user_addr = NULL, *phys_addrs = NULL, cma_mem_size = 0;
-	uint64_t rounddown_num = rounddown_pow_of_two(cma_info->entry_num);
-
-	BUG_ON(!mutex_is_locked(&drm_p->mm_lock));
-
-	if (!num)
-		return -ENODEV;
-	/* Limited by hardware, the entry number can only be power of 2
-	 * rounddown_pow_of_two 255=>>128 63=>>32
-	 */
-	if (rounddown_num != cma_info->entry_num) {
-		DRM_ERROR("Request %lld, round down to power of 2 %lld\n", 
-				cma_info->entry_num, rounddown_num);
-		return -EINVAL;
-	}
-
-	if (rounddown_num > num)
-		return -EINVAL;
-
-	user_addr = vzalloc(sizeof(uint64_t)*rounddown_num);
-	if (!user_addr)
-		return -ENOMEM;
-
-	ret = copy_from_user(user_addr, cma_info->user_addr, sizeof(uint64_t)*rounddown_num);
-	if (ret) {
-		ret = -EFAULT;
-		goto done;
-	}
-
-	for (i = 0; i < rounddown_num-1; ++i) {
-		for (j = i+1; j < rounddown_num; ++j) {
-			if (user_addr[i] == user_addr[j]) {
-				ret = -EINVAL;
-				DRM_ERROR("duplicated Huge Page");
-				goto done;
-			}
-		}
-	}
-
-	for (i = 0; i < rounddown_num; ++i) {
-		if (user_addr[i] & (page_sz - 1)) {
-			DRM_ERROR("Invalid Huge Page");
-			ret = -EINVAL;
-			goto done;
-		}
-
-		ret = xocl_cma_mem_alloc_huge_page_by_idx(drm_p, i, user_addr[i], page_sz);
-		if (ret)
-			goto done;
-	}
-
-	phys_addrs = vzalloc(rounddown_num*sizeof(uint64_t));
-	if (!phys_addrs) {
-		ret = -ENOMEM;
-		goto done;		
-	}
-
-	for (i = 0; i < rounddown_num; ++i) {
-		struct xocl_cma_memory *cma_mem = &drm_p->cma_bank->cma_mem[i];
-
-		if (!cma_mem) {
-			ret = -ENOMEM;
-			break;
-		}
-
-		/* All the cma mem should have the same size,
-		 * find the black sheep
-		 */
-		if (cma_mem_size && cma_mem_size != cma_mem->size) {
-			DRM_ERROR("CMA memory mixmatch");
-			ret = -EINVAL;
-			break;
-		}
-
-		phys_addrs[i] = cma_mem->paddr;
-		cma_mem_size = cma_mem->size;
-	}
-
-	if (ret)
-		goto done;
-
-	/* Remember how many cma mem we allocate*/
-	drm_p->cma_bank->entry_num = rounddown_num;
-	drm_p->cma_bank->entry_sz = page_sz;
-
-	ret = xocl_addr_translator_set_page_table(xdev, phys_addrs, page_sz, rounddown_num);
-done:
-	vfree(user_addr);
-	vfree(phys_addrs);
-	return ret;
-}
-
-static struct page **xocl_virt_addr_get_pages(void *vaddr, int npages)
-
-{
-	struct page *p, **pages;
-	int i;
-	uint64_t offset = 0;
-
-	pages = vzalloc(npages*sizeof(struct page *));
-	if (pages == NULL)
-		return ERR_PTR(-ENOMEM);
-
-	for (i = 0; i < npages; i++) {
-		p = virt_to_page(vaddr + offset);
-		pages[i] = p;
-		if (IS_ERR(p))
-			goto fail;
-		offset += PAGE_SIZE;
-	}
-
-	return pages;
-fail:
-	vfree(pages);
-	return ERR_CAST(p);
-}
-
-static int xocl_cma_mem_alloc_by_idx(struct xocl_drm *drm_p, uint64_t size, uint32_t idx)
-{
-	int ret = 0;
-	uint64_t page_count;
-	struct xocl_cma_memory *cma_mem = &drm_p->cma_bank->cma_mem[idx];
-	struct page **pages = NULL;
-	dma_addr_t dma_addr;
-
-	page_count = (size) >> PAGE_SHIFT;
-
-	cma_mem->vaddr = dma_alloc_coherent(&drm_p->ddev->pdev->dev, size, &dma_addr, GFP_KERNEL);
-
-	if (!cma_mem->vaddr) {
-		DRM_ERROR("Unable to alloc %llx bytes CMA buffer", size);
-		ret = -ENOMEM;
-		goto done;
-	}
-
-	pages = xocl_virt_addr_get_pages(cma_mem->vaddr, size >> PAGE_SHIFT);
-	if (IS_ERR(pages)) {
-		ret = PTR_ERR(pages);
-		goto done;
-	}
-
-	cma_mem->pages = pages;
-	cma_mem->paddr = dma_addr;
-	cma_mem->size = size;
-
-done:
-	if (ret)
-		xocl_cma_mem_free(drm_p, idx);
-
-	return ret;
-}
-
-static void __xocl_cma_bank_free(struct xocl_drm *drm_p)
-{
-	if (!drm_p->cma_bank)
-		return;
-
-	xocl_cma_mem_free_all(drm_p);
-	xocl_addr_translator_clean(drm_p->xdev);
-	vfree(drm_p->cma_bank);
-	drm_p->cma_bank = NULL;
-}
-
-static int xocl_cma_mem_alloc(struct xocl_drm *drm_p, uint64_t size)
-{
-	xdev_handle_t xdev = drm_p->xdev;
-	int ret = 0;
-	uint64_t i = 0, page_sz;
-	uint64_t page_num = xocl_addr_translator_get_entries_num(xdev);
-	uint64_t *phys_addrs = NULL, cma_mem_size = 0;
-
-	if (!page_num) {
-		DRM_ERROR("Doesn't support CMA BANK feature");
-		return -ENODEV;		
-	}
-
-	page_sz = size/page_num;
-
-	if (page_sz < PAGE_SIZE || !is_power_of_2(page_sz)) {
-		DRM_ERROR("Invalid CMA bank size");
-		return -EINVAL;
-	}
-
-	if (page_sz > (PAGE_SIZE << (MAX_ORDER-1))) {
-		DRM_WARN("Unable to allocate with page size 0x%llx", page_sz);
-		return -EINVAL;
-	}
-
-	for (; i < page_num; ++i) {
-		ret = xocl_cma_mem_alloc_by_idx(drm_p, page_sz, i);
-		if (ret) {
-			ret = -ENOMEM;
-			goto done;
-		}
-	}
-
-	phys_addrs = vzalloc(page_num*sizeof(uint64_t));
-	if (!phys_addrs) {
-		ret = -ENOMEM;
-		goto done;		
-	}
-
-	for (i = 0; i < page_num; ++i) {
-		struct xocl_cma_memory *cma_mem = &drm_p->cma_bank->cma_mem[i];
-
-		if (!cma_mem) {
-			ret = -ENOMEM;
-			break;
-		}
-
-		/* All the cma mem should have the same size,
-		 * find the black sheep
-		 */
-		if (cma_mem_size && cma_mem_size != cma_mem->size) {
-			DRM_ERROR("CMA memory mixmatch");
-			ret = -EINVAL;
-			break;
-		}
-
-		phys_addrs[i] = cma_mem->paddr;
-		cma_mem_size = cma_mem->size;
-	}
-
-	if (ret)
-		goto done;
-
-	drm_p->cma_bank->entry_num = page_num;
-	drm_p->cma_bank->entry_sz = page_sz;
-
-	ret = xocl_addr_translator_set_page_table(xdev, phys_addrs, page_sz, page_num);
-done:	
-	vfree(phys_addrs);
-	return ret;
-}
-
-void xocl_cma_bank_free(struct xocl_drm *drm_p)
-{
-	mutex_lock(&drm_p->mm_lock);
-	__xocl_cma_bank_free(drm_p);
-	xocl_cleanup_mem_nolock(drm_p);
-	xocl_icap_clean_bitstream(drm_p->xdev);
-	mutex_unlock(&drm_p->mm_lock);
-}
-
-int xocl_cma_bank_alloc(struct xocl_drm *drm_p, struct drm_xocl_alloc_cma_info *cma_info)
-{
-	int err = 0;
-	xdev_handle_t xdev = drm_p->xdev;
-	int num = xocl_addr_translator_get_entries_num(xdev);
-
-	if (!num) {
-		DRM_ERROR("Doesn't support HOST MEM feature");
-		return -ENODEV;
-	}
-
-	mutex_lock(&drm_p->mm_lock);
-
-	xocl_cleanup_mem_nolock(drm_p);
-	xocl_icap_clean_bitstream(drm_p->xdev);
-
-	if (drm_p->cma_bank) {
-		uint64_t allocated_size = drm_p->cma_bank->entry_num * drm_p->cma_bank->entry_sz;
-		if (allocated_size == cma_info->total_size) {
-			DRM_INFO("HOST MEM already allocated, skip");
-			goto unlock;
-		} else {
-			DRM_ERROR("HOST MEM already allocated, size 0x%llx", allocated_size);
-			DRM_ERROR("Please run xbutil host disable first");
-			err = -EBUSY;
-			goto unlock;
-		}
-	}
-
-	drm_p->cma_bank = vzalloc(sizeof(struct xocl_cma_bank)+num*sizeof(struct xocl_cma_memory));
-	if (!drm_p->cma_bank) {
-		err = -ENOMEM;
-		goto done;
-	}
-
-	if (cma_info->entry_num)
-		err = xocl_cma_mem_alloc_huge_page(drm_p, cma_info);
-	else {
-		/* Cast all err as E2BIG */
-		err = xocl_cma_mem_alloc(drm_p, cma_info->total_size);
-		if (err) {
-			err = -E2BIG;
-			goto done;
-		}
-	}
-done:
-	if (err)
-		__xocl_cma_bank_free(drm_p);
-unlock:
-	mutex_unlock(&drm_p->mm_lock);
-	DRM_INFO("%s, %d", __func__, err);
-	return err;
 }
