+++ src/runtime_src/core/pcie/driver/linux/xocl/userpf/xocl_ioctl.c
@@ -579,7 +579,7 @@ int xocl_alloc_cma_ioctl(struct drm_device *dev, void *data,
 		goto done;
 	}
 
-	err = xocl_cma_bank_alloc(drm_p, cma_info);
+	err = xocl_cma_bank_alloc(xdev, cma_info);
 done:
 	mutex_unlock(&xdev->dev_lock);
 	return err;
@@ -594,11 +594,10 @@ int xocl_free_cma_ioctl(struct drm_device *dev, void *data,
 
 	mutex_lock(&xdev->dev_lock);
 
-
 	if (xocl_xclbin_in_use(xdev) || xocl_check_topology(drm_p))
 		err = -EBUSY;
 	else
-		xocl_cma_bank_free(drm_p);
+		xocl_cma_bank_free(xdev);
 	mutex_unlock(&xdev->dev_lock);
 
 	return err;
