+++ src/runtime_src/core/pcie/driver/linux/xocl/xocl_drm.h
@@ -61,7 +61,6 @@ struct xocl_drm {
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(3, 7, 0)
 	DECLARE_HASHTABLE(mm_range, 6);
 #endif
-	struct xocl_cma_bank  *cma_bank;
 };
 
 struct drm_xocl_bo {
@@ -106,8 +105,6 @@ int xocl_init_mem(struct xocl_drm *drm_p);
 int xocl_cleanup_mem(struct xocl_drm *drm_p);
 
 bool is_cma_bank(struct xocl_drm *drm_p, uint32_t memidx);
-int xocl_cma_bank_alloc(struct xocl_drm *drm_p, struct drm_xocl_alloc_cma_info *cma_info);
-void xocl_cma_bank_free(struct xocl_drm *drm_p);
 int xocl_check_topology(struct xocl_drm *drm_p);
 
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(4, 17, 0)
